Commit #1.5 (Still Teleports)
Added a new command that should work better in terms of grabbing rings. Also changed the backup to be in the proper place, though it can still be used normally. Also, the auto teleports during simulation (A feat of modern engineering and other stuff, if I do say so myself, which I do!) Let's hope it works when the robot actually exists again. Note how the command change uses a fake subsystem within the RobotContainer. It's funny how that works. Ok, now I'm finished ranting.