Commit #3(I think) Vision Processing
I'm extremely happy with this. The amp shoot hasn't changed, but the limelight 3 on our robot now should function entirely. The code isn't exactly clean, and some things could've been done better, but I'm happy with this for the moment. My hardest challenge was accessing some of the classes for tracking. I decided to contain most of the code required inside of Robot.java, since that means I can access any other class. I haven't tested, so this could be very sad, but we can hope! Going to test tomorrow hopefully.