Work On Commands For Vision
The hard limit didn't work during testing, and the bindings ran continually, due to an error. We are going to use an InterpolatingDoubleTreeMap to control the vision angle, based on the yOffset we have. You can set a number of values(the yOffset), and get back the angle we want.