Vision Preparation (Without Any Testing!)
I'm preparing for the next part of using vision, before finishing the part I'm already working on. Admittedly, this probably isn't the optimal way that I'm planning, but it should be good for what we need. Not everything has been added(I want to finish the speaker angle first before doing all that.), but this is a good foundation for what I plan to do once we figure out what angles should be mapped.