Documentation(along with fixes)
It's been a while since I worked on vision(nearly three months!!), and I thought the best way to remember everything would be to document how it all works at the moment. This will not only allow me to understand my own code, but will also help me when explaining to others my ideas of how this should work. I plan to do this for other portions of the robot code, but that's for later. I also fixed an issue with removing the limit, where the shooter would become immovable, whether vision is on or off. Any documentation I create will be inside a new "docs" folder.